Revert "Revert "ExtensionResource""
This reverts commit 844faeba1171ba77716938615cef44caab1fe137.
