+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/model/v2/ExtensionResource.java
@@ -7,12 +7,21 @@ import java.util.Map;
 import org.apache.commons.lang3.ObjectUtils;
 
 import com.sap.cloud.lm.sl.common.util.MapUtil;
+import com.sap.cloud.lm.sl.mta.model.ElementContext;
+import com.sap.cloud.lm.sl.mta.model.NamedElement;
 import com.sap.cloud.lm.sl.mta.model.ParametersContainer;
+import com.sap.cloud.lm.sl.mta.model.PropertiesContainer;
+import com.sap.cloud.lm.sl.mta.model.VisitableElement;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
 import com.sap.cloud.lm.sl.mta.parsers.v2.ExtensionResourceParser;
 import com.sap.cloud.lm.sl.mta.util.YamlElement;
 
-public class ExtensionResource extends com.sap.cloud.lm.sl.mta.model.v1.ExtensionResource implements ParametersContainer {
+public class ExtensionResource implements VisitableElement, NamedElement, PropertiesContainer, ParametersContainer {
 
+    @YamlElement(ExtensionResourceParser.NAME)
+    private String name;
+    @YamlElement(ExtensionResourceParser.PROPERTIES)
+    private Map<String, Object> properties;
     @YamlElement(ExtensionResourceParser.PARAMETERS)
     private Map<String, Object> parameters;
 
@@ -20,21 +29,44 @@ public class ExtensionResource extends com.sap.cloud.lm.sl.mta.model.v1.Extensio
 
     }
 
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public Map<String, Object> getProperties() {
+        return MapUtil.unmodifiable(properties);
+    }
+    
     @Override
     public Map<String, Object> getParameters() {
         return MapUtil.unmodifiable(parameters);
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setProperties(Map<String, Object> properties) {
+        this.properties = new LinkedHashMap<>(properties);
     }
 
     @Override
     public void setParameters(Map<String, Object> parameters) {
         this.parameters = new LinkedHashMap<>(parameters);
     }
+    
+    @Override
+    public void accept(ElementContext context, Visitor visitor) {
+        visitor.visit(context, this);
+    }
 
-    public static class Builder extends com.sap.cloud.lm.sl.mta.model.v1.ExtensionResource.Builder {
+    public static class Builder {
 
+        protected String name;
+        protected Map<String, Object> properties;
         protected Map<String, Object> parameters;
 
-        @Override
         public ExtensionResource build() {
             ExtensionResource result = new ExtensionResource();
             result.setName(name);
@@ -43,10 +75,18 @@ public class ExtensionResource extends com.sap.cloud.lm.sl.mta.model.v1.Extensio
             return result;
         }
 
+        public void setName(String name) {
+            this.name = name;
+        }
+
+        public void setProperties(Map<String, Object> properties) {
+            this.properties = properties;
+        }
+
         public void setParameters(Map<String, Object> parameters) {
             this.parameters = parameters;
         }
 
     }
-
+    
 }
